adding export of primers; fixed on complete workflow output; added max cpus default for index_genome;

+++ modules/local/gen_target_info_from_genomes/main.nf
@@ -7,10 +7,12 @@ process GEN_TARGET_INFO_FROM_GENOMES_NANOPORE {
     publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "genome_extraction/locationsByGenome"
     publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "genome_extraction/allExtractionCounts.tab.txt"
     publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "targets_with_extractins.txt"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "primers.tsv"
+
 
 
     input:
-    path primers_fnp
+    path primers_fnp, name : "primers.tsv"
     val pub_dir
     path genome_dir
     path gff_dir
@@ -22,6 +24,7 @@ process GEN_TARGET_INFO_FROM_GENOMES_NANOPORE {
     path "genome_extraction/locationsByGenome", emit: locations_by_genome
     path "genome_extraction/allExtractionCounts.tab.txt", emit: all_extraction_counts
     path "targets_with_extractins.txt", emit: targets_with_extractins
+    path "primers.tsv", emit: primers
 
     script:
     if (file("${pub_dir}/genome_extraction/forSeekDeep").exists() && file("${primers_fnp}").lastModified() < file("${pub_dir}/genome_extraction/forSeekDeep").lastModified()){
@@ -43,7 +46,17 @@ process GEN_TARGET_INFO_FROM_GENOMES_NANOPORE {
                 --errors ${errors_allowed} \
                 --numThreads ${ncpus} \
                 --useBlast
-        elucidator tableExtractCriteria --file genome_extraction/allExtractionCounts.tab.txt --delim tab --columnName extractionCounts --header --cutOff 0 | elucidator printCol --file STDIN --delim tab --header --columnName target --sort --unique  > targets_with_extractins.txt
+        elucidator tableExtractCriteria --file genome_extraction/allExtractionCounts.tab.txt \
+                --delim tab \
+                --columnName extractionCounts \
+                --header \
+                --cutOff 0 | elucidator printCol \
+                                        --file STDIN \
+                                        --delim tab \
+                                        --header \
+                                        --columnName target \
+                                        --sort \
+                                        --unique  > targets_with_extractins.txt
         """
     }
 }
@@ -53,9 +66,11 @@ process GEN_TARGET_INFO_FROM_GENOMES_ILLUMINA {
     cpus   { ncpus }
 
 
-    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "forSeekDeep"
-    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "locationsByGenome"
-    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "allExtractionCounts.tab.txt"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "genome_extraction/forSeekDeep"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "genome_extraction/locationsByGenome"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "genome_extraction/allExtractionCounts.tab.txt"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "targets_with_extractins.txt"
+    publishDir "${pub_dir}", mode: 'copy', overwrite: true, pattern: "${primers_fnp}", saveAs: "primers.tsv"
 
 
     input:
@@ -68,22 +83,44 @@ process GEN_TARGET_INFO_FROM_GENOMES_ILLUMINA {
     val ncpus
 
     output:
-    path "forSeekDeep", emit: for_seek_deep_info
-    path "locationsByGenome", emit: locations_by_genome
-    path "extraction/allExtractionCounts.tab.txt", emit: all_extraction_counts
+    path "genome_extraction/forSeekDeep", emit: for_seek_deep_info
+    path "genome_extraction/locationsByGenome", emit: locations_by_genome
+    path "genome_extraction/allExtractionCounts.tab.txt", emit: all_extraction_counts
+    path "targets_with_extractins.txt", emit: targets_with_extractins
 
     script:
+    if (file("${pub_dir}/genome_extraction/forSeekDeep").exists() && file("${primers_fnp}").lastModified() < file("${pub_dir}/genome_extraction/forSeekDeep").lastModified()){
+        """
+        mkdir genome_extraction
+        cp -r "${pub_dir}/genome_extraction/forSeekDeep" genome_extraction/
+        cp -r "${pub_dir}/genome_extraction/locationsByGenome" genome_extraction/
+        cp "${pub_dir}/genome_extraction/allExtractionCounts.tab.txt" genome_extraction/
+        elucidator tableExtractCriteria --file genome_extraction/allExtractionCounts.tab.txt --delim tab --columnName extractionCounts --header --cutOff 0 | elucidator printCol --file STDIN --delim tab --header --columnName target --sort --unique  > targets_with_extractins.txt
+        """
+    } else {
     """
-    SeekDeep genTargetInfoFromGenomes \
-            --primers ${primers_fnp} \
-            --pairedEndLength ${paired_end_length} \
-            --genomeDir ${genome_dir} \
-            --gffDir ${gff_dir} \
-            --dout extraction \
-            --errors ${errors_allowed} \
-            --numThreads ${ncpus} \
-            --useBlast
-    ln -s extraction/forSeekDeep
-    ln -s extraction/locationsByGenome
+        SeekDeep genTargetInfoFromGenomes \
+                --primers ${primers_fnp} \
+                --pairedEndLength ${paired_end_length} \
+                --genomeDir ${genome_dir} \
+                --gffDir ${gff_dir} \
+                --dout extraction \
+                --errors ${errors_allowed} \
+                --numThreads ${ncpus} \
+                --useBlast
+        ln -s extraction/forSeekDeep
+        ln -s extraction/locationsByGenome
+        elucidator tableExtractCriteria --file genome_extraction/allExtractionCounts.tab.txt \
+            --delim tab \
+            --columnName extractionCounts \
+            --header \
+            --cutOff 0 | elucidator printCol \
+                                    --file STDIN \
+                                    --delim tab \
+                                    --header \
+                                    --columnName target \
+                                    --sort \
+                                    --unique  > targets_with_extractins.txt
     """
+    }
 }
